+++ wcfsetup/install/files/lib/form/EmailActivationForm.class.php
@@ -108,7 +108,7 @@ final class EmailActivationForm extends AbstractFormBuilderForm
      */
     public function show()
     {
-        if (!(REGISTER_ACTIVATION_METHOD & User::REGISTER_ACTIVATION_USER)) {
+        if (!((int)REGISTER_ACTIVATION_METHOD & User::REGISTER_ACTIVATION_USER)) {
             throw new IllegalLinkException();
         }
