+++ wcfsetup/install/files/lib/form/RegisterNewActivationCodeForm.class.php
@@ -207,7 +207,7 @@ final class RegisterNewActivationCodeForm extends AbstractFormBuilderForm
      */
     public function show()
     {
-        if (!(REGISTER_ACTIVATION_METHOD & User::REGISTER_ACTIVATION_USER)) {
+        if (!((int)REGISTER_ACTIVATION_METHOD & User::REGISTER_ACTIVATION_USER)) {
             throw new IllegalLinkException();
         }
